refactor: share the upgrade-readiness aggregator as a service-layer bean
The upgradeReadiness actuator endpoint built its own
MigrationStatusAggregator instance internally, from a directly-injected
List<MigrationStatusProvider>. That aggregator is the only place that
folds every registered condition (RocksDB, ES/OS schema, RDBMS schema)
into a per-physical-tenant, per-condition status map, with caching that
keeps the last confirmed MIGRATED result and backfills gaps so a
transient provider failure doesn't regress a previously-confirmed
condition.

A second, public consumer needs that same fold (camunda/camunda#61619).
Duplicating the aggregation logic would risk the two endpoints silently
disagreeing whenever a provider hiccups, since each would keep its own
independent cache.

Move the aggregation engine out of dist and into the service module,
which sits between REST controllers and the engine and already hosts
the closest analog (ClusterStatusServices). Register it as the one
Spring bean both consumers will share, wired from a small, permanently
active configuration class (not gated behind any HTTP gateway, since
the actuator must keep working on broker-only nodes). aggregate() now
returns the raw per-tenant/per-condition map instead of the
actuator-specific UpgradeReadinessResponse; the endpoint derives its
upgradeable boolean from that map itself, keeping the shared engine
free of any one consumer's interpretation of the data.

No behavioral change: the endpoint was already a singleton Spring bean
building exactly one aggregator, so there was already only one instance
in practice.
