+++ service/src/test/java/io/camunda/service/MigrationStatusAggregatorTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.shared.management;
+package io.camunda.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -19,20 +19,19 @@ import org.junit.jupiter.api.Test;
 final class MigrationStatusAggregatorTest {
 
   @Test
-  void shouldReportNotUpgradeableWhenNoProviderIsRegistered() {
+  void shouldReportNoPhysicalTenantsWhenNoProviderIsRegistered() {
     // given - staged rollout: not every provider exists yet
     final var aggregator = new MigrationStatusAggregator(List.of());
 
     // when
-    final var response = aggregator.aggregate();
+    final var physicalTenants = aggregator.aggregate();
 
     // then - an empty condition set must never be mistaken for readiness
-    assertThat(response.upgradeable()).isFalse();
-    assertThat(response.physicalTenants()).isEmpty();
+    assertThat(physicalTenants).isEmpty();
   }
 
   @Test
-  void shouldReportUpgradeableOnlyWhenEveryConditionIsMigratedForEveryTenant() {
+  void shouldReportEveryConditionForEveryPhysicalTenant() {
     // given
     final var aggregator =
         new MigrationStatusAggregator(
@@ -41,11 +40,10 @@ final class MigrationStatusAggregatorTest {
                 provider("b", Map.of("default", migrated("b done")))));
 
     // when
-    final var response = aggregator.aggregate();
+    final var physicalTenants = aggregator.aggregate();
 
     // then
-    assertThat(response.upgradeable()).isTrue();
-    assertThat(response.physicalTenants().get("default")).hasSize(2);
+    assertThat(physicalTenants.get("default")).hasSize(2);
   }
 
   @Test
@@ -61,18 +59,16 @@ final class MigrationStatusAggregatorTest {
                         "tenantB", inProgress("b behind")))));
 
     // when
-    final var response = aggregator.aggregate();
+    final var physicalTenants = aggregator.aggregate();
 
     // then
-    assertThat(response.upgradeable()).isFalse();
-    assertThat(response.physicalTenants().get("tenantA").get("a").state())
-        .isEqualTo(MigrationState.MIGRATED);
-    assertThat(response.physicalTenants().get("tenantB").get("a").state())
+    assertThat(physicalTenants.get("tenantA").get("a").state()).isEqualTo(MigrationState.MIGRATED);
+    assertThat(physicalTenants.get("tenantB").get("a").state())
         .isEqualTo(MigrationState.MIGRATION_IN_PROGRESS);
   }
 
   @Test
-  void shouldReportNotUpgradeableWhenOneConditionIsNotMigratedForOneTenant() {
+  void shouldKeepEachConditionIndependentPerTenant() {
     // given
     final var aggregator =
         new MigrationStatusAggregator(
@@ -81,11 +77,11 @@ final class MigrationStatusAggregatorTest {
                 provider("b", Map.of("default", inProgress("b behind")))));
 
     // when
-    final var response = aggregator.aggregate();
+    final var physicalTenants = aggregator.aggregate();
 
     // then
-    assertThat(response.upgradeable()).isFalse();
-    assertThat(response.physicalTenants().get("default").get("b").state())
+    assertThat(physicalTenants.get("default").get("a").state()).isEqualTo(MigrationState.MIGRATED);
+    assertThat(physicalTenants.get("default").get("b").state())
         .isEqualTo(MigrationState.MIGRATION_IN_PROGRESS);
   }
 
@@ -95,12 +91,11 @@ final class MigrationStatusAggregatorTest {
     final var aggregator = new MigrationStatusAggregator(List.of(throwingProvider("a", "boom")));
 
     // when
-    final var response = aggregator.aggregate();
+    final var physicalTenants = aggregator.aggregate();
 
     // then - a broken provider must never crash the endpoint; with no known tenant, there's
     // simply nothing to report yet
-    assertThat(response.upgradeable()).isFalse();
-    assertThat(response.physicalTenants()).isEmpty();
+    assertThat(physicalTenants).isEmpty();
   }
 
   @Test
@@ -113,11 +108,10 @@ final class MigrationStatusAggregatorTest {
                 throwingProvider("b", "boom")));
 
     // when
-    final var response = aggregator.aggregate();
+    final var physicalTenants = aggregator.aggregate();
 
     // then - "b" must not silently disappear for "default": a gap is UNKNOWN, never omission
-    assertThat(response.upgradeable()).isFalse();
-    final var conditions = response.physicalTenants().get("default");
+    final var conditions = physicalTenants.get("default");
     assertThat(conditions.get("a").state()).isEqualTo(MigrationState.MIGRATED);
     assertThat(conditions.get("b").state()).isEqualTo(MigrationState.UNKNOWN);
     assertThat(conditions.get("b").detail()).contains("no status reported");
@@ -131,13 +125,13 @@ final class MigrationStatusAggregatorTest {
     final var aggregator = new MigrationStatusAggregator(List.of(flakyProvider));
 
     // when - first poll confirms MIGRATED, second poll would otherwise lose the entry entirely
-    final var firstResponse = aggregator.aggregate();
-    final var secondResponse = aggregator.aggregate();
+    final var firstPhysicalTenants = aggregator.aggregate();
+    final var secondPhysicalTenants = aggregator.aggregate();
 
     // then - monotonicity is preserved via the backfill, even under total provider failure
-    assertThat(firstResponse.physicalTenants().get("default").get("a").state())
+    assertThat(firstPhysicalTenants.get("default").get("a").state())
         .isEqualTo(MigrationState.MIGRATED);
-    assertThat(secondResponse.physicalTenants().get("default").get("a").state())
+    assertThat(secondPhysicalTenants.get("default").get("a").state())
         .isEqualTo(MigrationState.MIGRATED);
   }
 
@@ -149,10 +143,10 @@ final class MigrationStatusAggregatorTest {
 
     // when
     aggregator.aggregate();
-    final var secondResponse = aggregator.aggregate();
+    final var secondPhysicalTenants = aggregator.aggregate();
 
     // then - nothing was ever confirmed MIGRATED, so the backfill falls back to UNKNOWN
-    assertThat(secondResponse.physicalTenants().get("default").get("a").state())
+    assertThat(secondPhysicalTenants.get("default").get("a").state())
         .isEqualTo(MigrationState.UNKNOWN);
   }
